Bug#11752665: WINDOWS SERVER X64: SO MANY COMPILER WARNINGS
Patch #1.

This patch reduces the compiler warning count on Win64 by about 1400,
as reported by VS2013.

Note that no struct/class are grown in size.

+++ sql/sql_prepare.h
@@ -214,7 +214,7 @@ public:
     @sa Documentation for C API function
     mysql_field_count()
   */
-  ulong get_field_count() const
+  size_t get_field_count() const
   {
     return m_current_rset ? m_current_rset->get_field_count() : 0;
   }
